add-页面优化，添加权限控制

+++ tencloud/src/main/java/com/chenk/tencentcloud/controller/HomeController.java
@@ -5,11 +5,14 @@ import com.chenk.tencentcloud.pojo.FileInsertDTO;
 import com.chenk.tencentcloud.pojo.ResultPage;
 import com.chenk.tencentcloud.pojo.bean.FileBean;
 import com.chenk.tencentcloud.service.FileService;
+import com.chenk.tencentcloud.util.TokenUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
+import java.util.UUID;
 
 /**
  * @author: chenke
@@ -42,7 +45,9 @@ public class HomeController {
         fileBean.setType(filePath.substring(filePath.lastIndexOf(".") + 1, filePath.length()));
         fileBean.setStatus(1L);
         fileBean.setRemark(null);
-//        fileBean.setSource("public");
+        if (!TokenUtil.isLogin()) {
+            fileBean.setSource("public");
+        }
         fileBean.setOriginFileName(dto.getOriginFileName());
         boolean b = fileService.add(fileBean);
         return b ? "上传成功" : "上传失败";
@@ -50,7 +55,17 @@ public class HomeController {
 
     @GetMapping("/remove")
     public String insertToDB(@RequestParam("fileName") String fileName) {
+        if (!TokenUtil.isLogin()) {
+            return "您无删除权限";
+        }
         boolean b = fileService.removeByFileName(fileName);
         return b ? "删除成功" : "删除失败";
+    }
+
+    @GetMapping("/login")
+    public String login(@RequestParam("pwd") String pwd) {
+        String token = UUID.randomUUID().toString();
+        TokenUtil.setDefaultToken(token);
+        return Objects.equals(pwd, TokenUtil.DEFAULT_PWD) ? token : null;
     }
 }
